requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
attrs==25.3.0
biopython==1.85
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.2
cycler==0.12.1
filelock==3.18.0
fonttools==4.57.0
frozenlist==1.6.0
fsspec==2025.3.2
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
multidict==6.4.3
networkx==3.4.2
numpy==2.2.4
packaging==24.2
pillow==11.2.1
propcache==0.3.1
psutil==7.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sympy==1.13.1
threadpoolctl==3.6.0
torch==2.6.0
torch-geometric==2.6.1
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==2.4.0
yarl==1.20.0
